Make SelfAnalyzingGrid a subclass of Grid
Add test for the consistency of the emptyCells() override in
SelfAnalyzingGrid.

Replace copyOf(...) factory methods in UnmodifiableGrid and Grid
with constructors, because constructors can be used to make instances
of subclasses.

SelfAnalyzingGrid does not have a constructor that takes
UnmodifiableGrid, in order to have one with special behavior for
SelfAnalyzingGrid arguments, while not violating the Liksov
substitution principle. SelfAnalyzingGrid instances can be
constructed from UnmodifiableGrid instances with the
fromOrdinaryGrid() factory method.